app/requirements.txt
pymongo==4.6.1
flask==3.0.2
python-dotenv==1.0.1
graphene==3.3
fastapi==0.110.0
strawberry-graphql[fastapi]==0.219.2
